Create AppFormDate.vue
