+++ components/AppForm.vue
@@ -1,21 +1,7 @@
 <script lang="ts" setup>
-import {
-  parseISO,
-  subDays,
-  subYears
-} from 'date-fns'
+import { parseISO } from 'date-fns'
 
 const formStore = useFormStore()
-const now = new Date()
-const onClickTodayButton = () => {
-  formStore.date = formatISODate(now)
-}
-const onClickYesterdayButton = () => {
-  formStore.date = formatISODate(subDays(now, 1))
-}
-const onClickOneYearAgoButton = () => {
-  formStore.date = formatISODate(subYears(now, 1))
-}
 const onClickSearchButton = () => {
   const url = createTwitterSearchURL({
     user: formStore.user,
@@ -58,34 +44,7 @@ const onClickTwisaveSearchButton = () => {
 <template>
   <el-form label-position="top">
     <AppFormUser />
-    <el-form-item label="日付">
-      <el-date-picker
-        v-model="formStore.date"
-        type="date"
-        value-format="YYYY-MM-DD"
-        clearable
-      />
-    </el-form-item>
-    <el-form-item>
-      <el-button
-        size="small"
-        @click="onClickTodayButton"
-      >
-        今日
-      </el-button>
-      <el-button
-        size="small"
-        @click="onClickYesterdayButton"
-      >
-        昨日
-      </el-button>
-      <el-button
-        size="small"
-        @click="onClickOneYearAgoButton"
-      >
-        1年前
-      </el-button>
-    </el-form-item>
+    <AppFormDate />
     <el-form-item label="キーワード">
       <el-input
         v-model="formStore.keyword"
